requirements.txt
python-dotenv==0.18.0
Flask==1.1.2
sklearn